Remove pages service in module, put everything in regular service file
