+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/VisibleNavbarItemList/VisibleNavbarItemList.test.tsx
@@ -5,7 +5,7 @@ import {
   reorderNavbarItem,
   removeNavbarItem,
 } from '../../../../services/navbar';
-import { deletePage } from '../../../../services/pages';
+import { deletePage } from 'services/pages';
 import { NAVIGATION_PATH } from '../..';
 import navbarItems from 'hooks/fixtures/navbarItems';
 import clHistory from 'utils/cl-router/history';
